+++ src/Domain/Events/EventId.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * This file is part of slick/cqrs-tools
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\CQRSTools\Domain\Events;
+
+use Slick\CQRSTools\Domain\AggregateRootIdentifier;
+
+/**
+ * EventId
+ *
+ * @package Slick\CQRSTools\Domain\Events
+*/
+final class EventId extends AggregateRootIdentifier
+{
+
+}