+++ src/Domain/Exception/IdentifierCreationException.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * This file is part of slick/cqrs-tools
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\CQRSTools\Domain\Exception;
+
+use RuntimeException;
+use Slick\CQRSTools\Exception;
+
+/**
+ * Identifier Creation Exception
+ *
+ * @package Slick\CQRSTools\Domain\Exception
+ */
+final class IdentifierCreationException extends RuntimeException implements Exception
+{
+
+}