+++ src/Event.php
@@ -10,14 +10,24 @@
 namespace Slick\CQRSTools;
 
 use DateTimeImmutable;
+use JsonSerializable;
+use League\Event\EventInterface;
+use Slick\CQRSTools\Domain\AggregateRootIdentifier;
+use Slick\CQRSTools\Domain\Events\EventId;
 
 /**
  * Interface Event
  *
  * @package Slick\CQRSTools
  */
-interface Event
+interface Event extends EventInterface, JsonSerializable
 {
+    /**
+     * Event identifier
+     *
+     * @return EventId
+     */
+    public function eventId(): EventId;
 
     /**
      * Date and time event has occurred
@@ -25,4 +35,11 @@ interface Event
      * @return DateTimeImmutable
      */
     public function occurredOn(): DateTimeImmutable;
+
+    /**
+     * The aggregate identifier of the entity responsible for this event
+     *
+     * @return null|AggregateRootIdentifier
+     */
+    public function author(): ?AggregateRootIdentifier;
 }