+++ src/Event/EventListener.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * This file is part of slick/cqrs-tools
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\CQRSTools\Event;
+
+use League\Event\ListenerInterface;
+
+/**
+ * Event Listener
+ *
+ * @package Slick\CQRSTools\Event
+ */
+interface EventListener extends ListenerInterface
+{
+
+}