+++ src/Exception/FailToCreateDateException.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ *This file is part of slick/cqrs-tools
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\CQRSTools\Exception;
+
+use RuntimeException;
+use Slick\CQRSTools\Exception;
+
+/**
+ * Fail To Create Date Exception
+ *
+ * @package Slick\CQRSTools\Exception
+ */
+class FailToCreateDateException extends RuntimeException implements Exception
+{
+
+}