+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Event` interface
+- Event generation and publishing system
+- Common domain tools: `Stringable` and `Comparable` interfaces
+- `AggregateRootIdentifier` abstract class
 
 ## [v0.1.0] 2018-08-20
 ### Added